+++ src/csulb/cecs327/Main.java
@@ -1,12 +1,17 @@
 package csulb.cecs327;
 
+import csulb.cecs327.Controllers.FrontEnd.AppUI;
+import csulb.cecs327.Models.User;
+
 import javax.swing.*;
 
 public class Main {
 
     public static void main(String[] args) {
         JFrame login = new JFrame("LoginPage");
-        login.setContentPane(new LoginPage());
+        // Dummy user
+        User user = new User("admin", "password", "email");
+        login.setContentPane(new AppUI(user));
         login.setSize(1024, 720);
         login.setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE);
         login.setVisible(true);
