+++ src/csulb/cecs327/Models/User.java
@@ -1,4 +1,4 @@
-package csulb.cecs327;
+package csulb.cecs327.Models;
 
 import java.util.HashMap;
