+++ src/csulb/cecs327/Services/UserSerializer.java
@@ -1,6 +1,7 @@
-package csulb.cecs327;
+package csulb.cecs327.Services;
 
 import com.google.gson.*;
+import csulb.cecs327.Models.User;
 
 public class UserSerializer implements JsonSerializer<User> {
     public JsonElement serialize(User user, java.lang.reflect.Type type, JsonSerializationContext jsonSerializationContext) {
